% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ac_measures_from_em.R
\name{ac_measures_from_em}
\alias{ac_measures_from_em}
\title{Calculate accuracy measures from error matrix}
\usage{
ac_measures_from_em(em, em_dimnames = NULL)
}
\arguments{
\item{em}{A confusion matrix (n x n)}

\item{em_dimnames}{Character vector of class names (n)}
}
\value{
A list given with the overall accuracy and per class users's and producers's accuracy
}
\description{
Calculate accuracy measures from error matrix
}
\examples{
error_m <- matrix(c(99, 2, 23, 45), ncol = 2)
ac_ac_measures_from_em(error_m, c("a", "b"))
}

