% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.null.class.R
\name{add.null.class}
\alias{add.null.class}
\title{Add a Null Class}
\usage{
add.null.class(target, predictors, nMC = 50)
}
\arguments{
\item{target}{Column with names of classes.}

\item{predictors}{The columns containig the predictor variables.}

\item{nMC}{Number returned observations of the null class (default = 50)}
}
\value{
Dataframe with original observatios and nMC observations of a null class.
First column named 'class' contains the target classes, predictor names are mantained.
}
\description{
Create nMC observations of a null class.
}
\details{
A null class has no multivariate structure. It is created by shuffling the 
rows across the whole dataset and then shuffling the column labels.
By doing so we generate nMC in silico observations for a class containing
only random variable values.
}
\examples{
data(iris)
iris.null <- add.null.class(iris$Species,iris[,1:4])
summary(iris.null)
}
\seealso{
\code{\link{smooth.data}}
}
\author{
Pedro Martinez Arbizu & Sven Rossel
}
