% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2xts.R
\name{df2xts}
\alias{df2xts}
\title{Convert dataframe to xts object}
\usage{
df2xts(x)
}
\arguments{
\item{x}{Dataframe to be converted}
}
\description{
Convert dataframe to xts object
}
\examples{
df2xts()
}
\keyword{dataframe}
\keyword{xts}
