% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi.R
\docType{data}
\name{maldi}
\alias{maldi}
\title{Example data maldi.RData for RFtools}
\format{
A data frame with 176 rows (specimens), 543 variables(protein masses) and a grouping factor (species)
}
\usage{
data(maldi)
}
\description{
This dataset is the intensity matrix generated by MALDI-TOF MS used in Rossel and Martínez Arbizu (submitted)
The Data shows MALDI-TOF MS spectra of 173 harpacticoid copepod specimens belonging to 16 different species from 
the North Sea benthic environment resulting in 176 different mass spectra. Three specimens were cut into to parts before 
measurement, accordingly resulting in two mass spectra per specimen. Data was generated using a Microflex LT/SH System 
(Bruker Daltonics) with a laser intensity between 50 to 60% in a mass range from 2000 to 20000.
Mass spectra were measured with flexControl 3.4 software (factory delivered Bruker Daltonics software).
Data was processed in R using packages MALDIquant and MALDIquantForeign, trimming mass spectra from 3000 to 15000 m/z.   
Spectra were smoothed (Savitzky & Golay), baseline removed (SNIP baseline estimation method) and spectra normalized using
TIC method. Noise was estimated using a signal to noise ratio of 7. Subsequently peaks were binned once as implemented in 
MALDIquant with a tolerance of 0.002 in a strict approach. The resulting intensity matrix was hellinger transformed (square root of
relative intensities) and is provided herewith.
Data was generated at Senckenberg am Meer Proteome Laboratory, German Center for Marine Biodiversity Research (DZMB), Wilhelmshaven, Germany
}
\details{
\itemize{
  \item code  Individual specimen code
  \item species  Name of the harpacticoid species
  \item Columns 3 to 544: m/z bins of MALDI-TOF spectrum showing hellinger transformed intensity 
   }
}
\author{
Sven Rossel  \email{sven.rossel@senckenberg.de}
}
\keyword{MALDI-TOF}
\keyword{Mass_spectra}
\keyword{data}
\keyword{datasets}
