% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSF.test.R
\name{MVSF.test}
\alias{MVSF.test}
\title{Multivariate Structure and Flexibility Test}
\usage{
MVSF.test(target, predictors, nMC = 999)
}
\arguments{
\item{target}{Vector or target Column with names of classes.}

\item{predictors}{Dataframe or matrix with predictor variables.}

\item{nMC}{Number of smoothed observations returned per class (default = 999)}
}
\value{
An object of class MVSF.
 Dataframe with results of multivariate test and flexibility test.
\item{oob.err}{OOB of trained model} 
\item{null.err}{mean OOB of null model} 
\item{P.null}{Pseudo P value, probability that trained OOB can happen by random} 
\item{q05.null}{0.5 quantile of the ecdf of the null error} 
\item{q95.null}{0.95 quantile of the ecdf of the null error} 
\item{smooth.err}{prediction error rate of the smoothed classes} 
\item{p.smooth}{p value of prop.test, probability that the trained error and the prediction error of smoothed classes are the same}
}
\description{
Test the multivariate structure of your data against an empirical null model
and the flexibility of your model to correctly classify new observations created by smoothing the classes.
}
\details{
Multivariate Test:
The observed OOB is compared against an empirical null model with no multivariate structure.
The null model is created by shuffling the lables of the classes nMC times
and keeping the OOB of each class under the null model. A Pseudo P-value is returned as P = r/nMC-1,
with r = the number of times that the null model OOB >= trained OOB, and nMC = number of Monte Carlo simulations.

Flexibility test:
Create nMC of each class observations smoothing the classes and predict using the trained model.
Return error rate as n/nMC, with n = total number of missclassifications and nMC = number of Monte Carlo simulations.       
prop.test is used for testing the null that trained OOB error = error of the smoothed observations

A plot method is available
}
\examples{
data(iris)
MVSF.iris <- MVSF.test(iris$Species,iris[,1:4],nMC=99)
print(MVSF.iris)
plot(MVSF.iris)
}
\seealso{
\code{\link{plot.MVSF}} \code{\link{robust.test}}
}
\author{
Pedro Martinez Arbizu & Sven Rossel
}
