% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MVSF.R
\name{plot.MVSF}
\alias{plot.MVSF}
\title{Plot Results of MVSF Tests}
\usage{
\method{plot}{MVSF}(x, pnull = 0.05, psmooth = 0.05)
}
\arguments{
\item{x}{Object created by MVSF.test()}

\item{pnull}{Significance level for the multivariate structure test}

\item{psmooth}{Significance level for the flexibility test}
}
\description{
Plot results of Multivariate Structure and Flexibility Test.
}
\details{
For each class, the trained OOB, the prediction error of smoothed class
and the error under null model is plotted. For the null error the 0.5-0.95 quantile interval is plotted as error line.
Filled symbols indicate significant difference under alfa level defined by pnull and psmooth.
}
\examples{
data(iris)
MVSF.iris <- MVSF.test(iris$Species,iris[,1:4])
print(MVSF.iris)
plot(MVSF.iris)
}
\author{
Pedro Martinez Arbizu
}
