% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RFPH.R
\name{plot.RFPH}
\alias{plot.RFPH}
\title{Plot Post-hoc test}
\usage{
\method{plot}{RFPH}(obj, Species = "all", q = 0.05)
}
\arguments{
\item{obj}{Object of class RFPH.}

\item{Species}{Name of class to be plotted or 'all' for all classes}

\item{q}{Significance level.}
}
\value{
Plot
}
\description{
Plots the results of rf.post.hoc.
}
\details{
If name of class to be plotted is specify, the plot returns
the density function of the probability of assignment to the correct class 
in the trainig data and the POA of the predicted data as circles.
Circle is filled 'red' if not significantly different at q level, otherwise 'white'.
The POA at significance level q is plotted as dashed grey line.

If Species = 'all' (default), then all classes are plotted. Instead of density function for trained POA, the median (grey symbol) 
and the range between quantiles between q and 1-q are plotted.
}
\examples{
data(iris)
library(randomForest)
irSe <- iris[iris$Species == 'setosa',]
ir <- iris[iris$Species != 'setosa',]
ir$Species <- factor(ir$Species)
irNc <- add.null.class(ir$Species,ir[,1:4])
rf <- randomForest(class ~ ., data = irNc)
posth <- rf.post.hoc(rf,irSe)

plot(posth)
plot(posth,'versicolor',0.001)
}
\seealso{
\code{\link{rf.post.hoc}}
}
\author{
Pedro Martinez Arbizu & Sven Rossel
}
