% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.votes_tsne.R
\name{plot.votes_tsne}
\alias{plot.votes_tsne}
\title{Plot votes_tsne}
\usage{
\method{plot}{votes_tsne}(
  x,
  cex = 1.5,
  pch = coltable$symbol[match(x$predicted, coltable$class)],
  bg = coltable$color[match(x$predicted, coltable$class)],
  main = "",
  alpha = 220,
  margin = 7,
  ncol = 4,
  x.intersp = 1,
  y.low = 3,
  pal = colors
)
}
\arguments{
\item{x}{A votes_tsne object}

\item{bg}{Fill color of symbols}

\item{main}{Title passed to plot}

\item{alpha}{The transparency value for the symbols}

\item{margin}{The Graphical parameter passed to oma for the lower margin, to accomodate legend}

\item{ncol}{The number of columns for the legend}

\item{x.intersp}{Space between legend symbol and legend text}

\item{y.low}{Distance of legend from y axis}

\item{pal}{Color palette to be sued for symbols (default is internal colors)}
}
\description{
Plot a votes_tsne object
}
\details{
Plot the resuls of a t-sne analysis of the randomForest votes matrix.
The fuction has an internal color palette 'colors' that allows for ajustement
of the transparency values using the parameter alpha.
USe your own color palette with the parameter pal (note that alpha will not wrok then)
The graph legend can be easyly placed at the bottom of the graph
using the parameters margin, ncol and y.low.
}
\examples{
data(maldi)
rf <- randomForest(species ~.,data = maldi[,-1])
vt <- votes.tsne(rf)
plot(vt)
plot(vt,margin=5,y.low=3,ncol=3,cex=2,alpha=180,main='rf votes on Maldi-Tof spectra')

}
\seealso{
\code{\link{votes.tsne}}
}
\author{
Pedro Martinez Arbizu
}
