% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.post.hoc.R
\name{rf.post.hoc}
\alias{rf.post.hoc}
\title{RF Post-Hoc Test}
\usage{
rf.post.hoc(rf, newdata)
}
\arguments{
\item{rf}{Object of class randomForest.}

\item{newdata}{Dataframe to predict using RF model.}
}
\value{
An object of class RF.ph List with results of Post-Hoc test .
\itemize{
\item{ecdf}{ The empirical cumulative distribution fuction of the beta distributed
 RF probability of assignment to correct class. }
\item{post.hoc}{ Dataframe with results of post-hoc test. }
 \itemize{
     \item{winner}{ The predicted class. }
     \item{POA}{ Probability of assigment to the winner class. }
     \item{p.post.hoc}{ Probability that POA belongs to POA distribution of this class in the trained model. }
     \item{sig}{ Significance code:  0 ***, 0.001 **, 0.01 *, 0.05 ., > ns  }
 }
}
}
\description{
Test for false positives in RF prediction.
}
\details{
RF Post-Hoc test:
Take an object created with randomForest and a new dataset to predict.
From the votes matrix of RF object take only correct assigments and calculate for each predicted class
the beta distribution of the probability of assignment (POA) to the correct class using fitdistr() from MASS.
In case fitdistr fails in estimating the parameters of beta,
the non-parametric empirical cumulative distribution function (ecdf) is used instead.
The for each observation in newdata, the probability that the POA to winner class (p.assign) belongs to
the beta (or ecdf) of that class in the training data set is calculated (p.post.hoc).
Use p.post.hoc to reject the hypothesis that the predicted POA belongs to the trained POA.
A low p.post.hoc probability (e.g. <= 0.05) indicates a possible missclassification of that observation.

Methods plot, print and summary are available
}
\examples{
#example with maldi data
data(maldi)
library(randomForest)
unique(maldi$species)
maldi_train <- maldi[maldi$species != 'Cletodes limicola',]
maldi_test <- maldi[maldi$species == 'Cletodes limicola',]
#exclude Cletodes limicola from factors
maldi_train$species <- factor(maldi_train$species)
rf <- randomForest(species ~ ., data = maldi_train[-1])
ph <- rf.post.hoc(rf,maldi_test)
plot(ph)
plot(ph,'Tachidius discipes')
}
\seealso{
\code{\link{add.null.class}} \code{\link{smooth.data}} \code{\link{robust.test}} \code{\link{plot.RFPH}}
}
\author{
Pedro Martinez Arbizu & Sven Rossel
}
