% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfcv_ncov.R
\name{rfcv_ncov}
\alias{rfcv_ncov}
\title{Title
A modified \code{\link[randomForest]{rfcv}} that returns the rf with he optimal ncov and the selected variables}
\usage{
rfcv_ncov(trainx, trainy, cv.fold = 5, scale = "log", step = 0.5,
  mtry = function(p) max(1, floor(sqrt(p))), recursive = FALSE,
  ntree = 500, ncore = parallel::detectCores(), verbose = TRUE, ...)
}
\arguments{
\item{trainx}{matrix or data frame containing columns of predictor
  variables}

\item{trainy}{vector of response, must have length equal to the number
  of rows in \code{trainx}}

\item{cv.fold}{number of folds in the cross-validation}

\item{scale}{if \code{"log"}, reduce a fixed proportion (\code{step})
  of variables at each step, otherwise reduce \code{step} variables at a
  time}

\item{step}{if \code{log=TRUE}, the fraction of variables to remove at
	each step, else remove this many variables at a time}

\item{mtry}{a function of number of remaining predictor variables to
	use as the \code{mtry} parameter in the \code{randomForest} call}

\item{recursive}{whether variable importance is (re-)assessed at each
  step of variable reduction}

\item{ncore}{Number of cores to use (set to 1 for sequential computing)}

\item{verbose}{Should progress indication be printed?}

\item{...}{additional arguments to \code{\link[randomForest]{randomForest}}}
}
\value{
A list with
\itemize{
 \item The optimized rf object
 \item A data.frame with the OOB and CV performance measures for the optimized and rf model
     using all variables (Regression) or the OOB and CV accuracy assesment results (Classification)
 \item A character vector with the names of the selected variables.
}
}
\description{
Title
A modified \code{\link[randomForest]{rfcv}} that returns the rf with he optimal ncov and the selected variables
}
\examples{
#' ## Classification:
##data(iris)
set.seed(71)
iris_rfncov <- rfcv_ncov(trainx=iris[ ,names(iris) != "Species"], trainy = iris$Species)
print(iris_rfncov)
}

