% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.data.R
\name{smooth.data}
\alias{smooth.data}
\title{Smooth Classes}
\usage{
smooth.data(target, predictors, nMC = 100)
}
\arguments{
\item{target}{Vector or target Column with names of classes.}

\item{predictors}{Dataframe or matrix with predictor variables.}

\item{nMC}{Number of smoothed observations returned per class (default = 100)}
}
\value{
Dataframe with nMC smoothed observations.
First column named 'class' contains the target classes, predictor names are mantained.
}
\description{
Create a new dataset with smoothed classes and nMC observations per class.
}
\details{
For each class, the function will shuffle the rows of each predictor separately
and extract nMC rows for each class. By doing so we generate nMC in silico observations for each class,
but mantaining the range (observed variability) for each predictor.
}
\examples{
data(iris)
iris.sm <- smooth.data(iris$Species,iris[,1:4])
summary(iris.sm)
}
\seealso{
\code{\link{add.null.class}}
}
\author{
Pedro Martinez Arbizu & Sven Rossel
}
