% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts2df.R
\name{xts2df}
\alias{xts2df}
\title{Convert xts to dataframe object}
\usage{
xts2df(x, date2RowName)
}
\arguments{
\item{x}{xts to be converted}

\item{date2RowName}{Do you want to covert the date to the rownames?}
}
\description{
Convert xts to dataframe object
}
\examples{
xts2df()
}
\keyword{dataframe}
\keyword{xts}
